requirements.txt
streamlit==1.4.0
Pillow>=8.1.1
seaborn==0.11.1
matplotlib==3.3.3
sklearn==0.0
scipy==1.5.4
numpy==1.20.0rc1
opencv-python-headless